#280: Corrections to shipping cost when inclusive of tax
If prices are displayed with tax, the shipping cost already includes its associated tax and that's the value that is carried forward as the order's shipping cost (used by ot_shipping).

This commit also contains some staging of corrections for #275, keeping some modicum of sanity on the initial recording of an order's shipping-tax-rate.